+++ System/Allocator.h
@@ -13,17 +13,17 @@ namespace System {
 
     class REGULAR_API IAllocator {
     public:
-        virtual Core::Expected<void *, System::AllocationError> Allocate(size_t size, bool initialize) = 0;
+        virtual Core::Expected<void *, System::AllocationError> Allocate(unsigned long size, bool initialize) = 0;
 
         virtual Core::Expected<Core::Empty, System::AllocationError> Deallocate(void *ptr) = 0;
 
-        virtual Core::Expected<void *, System::AllocationError> Resize(void *ptr, size_t size, bool initialize) = 0;
+        virtual Core::Expected<void *, System::AllocationError> Resize(void *ptr, unsigned long size, bool initialize) = 0;
 
-        virtual Core::Expected<Core::Empty, System::AllocationError> Copy(void *sourcePointer, void *targetPointer, size_t size) = 0;
+        virtual Core::Expected<Core::Empty, System::AllocationError> Copy(void *sourcePointer, void *targetPointer, unsigned long size) = 0;
 
-        virtual Core::Expected<Core::Empty, System::AllocationError> Move(void *sourcePointer, void *targetPointer, size_t size) = 0;
+        virtual Core::Expected<Core::Empty, System::AllocationError> Move(void *sourcePointer, void *targetPointer, unsigned long size) = 0;
 
-        virtual Core::Expected<Core::Empty, System::AllocationError> Zero(void *ptr, size_t size) = 0;
+        virtual Core::Expected<Core::Empty, System::AllocationError> Zero(void *ptr, unsigned long size) = 0;
 
 
     };
@@ -33,17 +33,17 @@ namespace System {
     public:
         Allocator();
 
-        Core::Expected<void *, System::AllocationError> Allocate(size_t size, bool initialize) noexcept;
+        Core::Expected<void *, System::AllocationError> Allocate(unsigned long size, bool initialize) noexcept;
 
         Core::Expected<Core::Empty, System::AllocationError> Deallocate(void *ptr) noexcept;
 
-        Core::Expected<void *, System::AllocationError> Resize(void *ptr, size_t size, bool initialize) noexcept;
+        Core::Expected<void *, System::AllocationError> Resize(void *ptr, unsigned long size, bool initialize) noexcept;
 
-        Core::Expected<Core::Empty, System::AllocationError> Copy(void *sourcePointer, void *targetPointer, size_t size) noexcept;
+        Core::Expected<Core::Empty, System::AllocationError> Copy(void *sourcePointer, void *targetPointer, unsigned long size) noexcept;
 
-        Core::Expected<Core::Empty, System::AllocationError> Move(void *sourcePointer, void *targetPointer, size_t size) noexcept;
+        Core::Expected<Core::Empty, System::AllocationError> Move(void *sourcePointer, void *targetPointer, unsigned long size) noexcept;
 
-        Core::Expected<Core::Empty, System::AllocationError> Zero(void *ptr, size_t size) noexcept;
+        Core::Expected<Core::Empty, System::AllocationError> Zero(void *ptr, unsigned long size) noexcept;
     };
 
 } // System
